==> markdowns/07_untreated_pca.RMd <==
---
title: "Untreated sample PCA"
author: "Elisha Roberson"
date: "2020-10-14"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Project - Untreated + Skin + Muscle

```{r load_libraries}
library( here )
library( cowplot )
library( tidyverse )
library( ggrepel )
```

```{r source}
source( file = here::here( "src", "shared_r_functions.R" ) )
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	fig.path = paste0( figure_dir, "/" ),
	fig.keep = 'all',
	dpi = 300,
	fig.width = 11,
	fig.height = 8.5
)
```

```{r constants}
fc_cutoff <- 2.0
```

```{r load_gene_lists}
untreated_v_control <- read_tsv( file = here( 'results', 'DiffExp_PBMC_Untreated_over_Control.tsv' ) ) %>%
	filter( qval < 0.05 & abs( FoldChange ) >= fc_cutoff ) %>%
	pull( gene_id )

status_de_gene_ids <- untreated_v_control
```

```{r gene_annotation}
ensembl <- read_tsv( file = here::here( 'info', 'biomart_gene_transcript_map.txt.gz' ), col_names = TRUE ) %>%
  dplyr::rename( .data = ., gene_id = `Gene stable ID` ) %>%
  dplyr::rename( .data = ., symbol = `Gene name` ) %>%
  dplyr::rename( .data = ., gene_biotype = `Gene type` ) %>%
  select( .data = ., gene_id, symbol, gene_biotype ) %>%
  as.data.frame( . ) %>%
  unique( . )

rownames( ensembl ) = pull( ensembl, gene_id )

ensembl_gene_ids <- pull( ensembl, gene_id )
```

```{r rgsm}
info <- read_csv( file = here::here( "info", "JDM_RNASeq_ReadGroup_Info.csv" ) ) %>%
  filter( Status %in% c( 'Control_PBMC', 'JDM_Untreated_PBMC', 'JDM_Inactive_PBMC' ) ) %>%
  select( -RGID ) %>%
  as.data.frame( . ) %>%
  unique( . ) %>%
  mutate( Status = case_when(
    str_detect( string = Status, pattern = "Control_PBMC" ) ~ "Control",
    str_detect( string = Status, pattern = "Untreated" ) ~ "JDM - Untreated",
    str_detect( string = Status, pattern = "Inactive" ) ~ "JDM - Inactive",
    TRUE ~ "Error"
  ) ) %>%
  mutate( Status = factor( Status, levels = c( 'Control', 'JDM - Untreated', 'JDM - Inactive' ) ) ) %>%
  as.data.frame( . )
rownames( info ) = info$RGSM
```

```{r rlog}
rlog <- read_tsv( file = here( 'results', 'unspliced_rlog_expression.tsv.gz' ), col_names = TRUE, col_types = cols( .default = 'd', gene_id = 'c' ) ) %>%
	as.data.frame( . ) %>%
	column_to_rownames( "gene_id" ) %>%
	as.matrix( . )

rlog <- rlog[ status_de_gene_ids, rownames( info ) ]
rownames( rlog ) = ensembl[ rownames( rlog ), "symbol" ]
```

```{r disease_status_pca_both_gene_sets}
first_plot_pc <- 1
second_plot_pc <- 2

# Run PCA
pca_out <- prcomp( x = t( rlog ), scale. = TRUE )
percent_variance <- round( pca_out$sdev^2 / sum( pca_out$sdev^2 )* 100.0, 2 )

# Add demographics
pca_coords <- pca_out$x %>%
	as.data.frame( . ) %>%
	rownames_to_column( ., "RGSM" ) %>%
	merge( info, ., by = "RGSM" )

# Plot dimension names
x_dim_name <- paste0( "PC", first_plot_pc, " (", percent_variance[first_plot_pc], "%)" )
y_dim_name <- paste0( "PC", second_plot_pc, " (", percent_variance[second_plot_pc], "%)" )

pca_ggplot <- ggplot( data = pca_coords, mapping = aes( x=PC1, y=PC2, shape=Status, colour=Status ) ) +
	theme_bw() +
  scale_color_manual( values = c( 'black', reds3Palette[3], reds3Palette[2] ) ) +
	geom_point( size=5 ) +
	gg_bigger_texts +
	gg_center_title +
	xlab( x_dim_name ) +
	ylab( y_dim_name ) #+
	#ggtitle( "Disease status classification by PCA" )

pca_ggplot
```

```{r loadings}
loadings_pc1 <- pca_out$rotation[,1]

load_df1 <- data.frame( gene_id = names( loadings_pc1 ), loading = unname( loadings_pc1 ) ) %>%
  dplyr::rename( symbol = gene_id ) %>%
  mutate( abs_loading = abs( loading ) ) %>%
  mutate( direction = case_when(
    loading < 0 ~ "Down",
    loading > 0 ~ "Up",
    TRUE ~ "Problem"
  )) %>%
  arrange( desc( abs_loading ) ) %>%
  .[ 1:20, ] %>%
  arrange( abs_loading ) %>%
  mutate( symbol = fct_inorder( symbol ) )

loadings1_plot <- ggplot( data = load_df1, mapping = aes( x = symbol, y = loading ) ) +
  geom_bar( stat = 'identity' ) +
  theme_bw() +
  xlab( "" ) + 
  ylab( "Loading" ) +
  ggtitle( "PC1 Loadings" ) +
  gg_center_title +
  coord_flip() +
  theme( axis.text.y = element_text( size = 10 ),
         axis.title = element_text( size = 10 ) )

############

loadings_pc2 <- pca_out$rotation[,2]

load_df2 <- data.frame( gene_id = names( loadings_pc2 ), loading = unname( loadings_pc2 ) ) %>%
  dplyr::rename( symbol = gene_id ) %>%
  mutate( abs_loading = abs( loading ) ) %>%
  mutate( direction = case_when(
    loading < 0 ~ "Down",
    loading > 0 ~ "Up",
    TRUE ~ "Problem"
  )) %>%
  arrange( desc( abs_loading ) ) %>%
  .[ 1:20, ] %>%
  arrange( abs_loading ) %>%
  mutate( symbol = fct_inorder( symbol ) )

loadings2_plot <- ggplot( data = load_df2, mapping = aes( x = symbol, y = loading ) ) +
  geom_bar( stat = 'identity' ) +
  theme_bw() +
  xlab( "" ) + 
  ylab( "Loading" ) +
  ggtitle( "PC2 Loadings" ) +
  gg_center_title +
  coord_flip() +
  theme( axis.text.y = element_text( size = 10 ),
         axis.title = element_text( size = 10 ) )
```

```{r pca_with_loadings}
left_panel <- plot_grid( pca_ggplot + theme( legend.position = 'top' ), labels = 'a' )
left_panel

right_panel <- plot_grid( loadings1_plot, loadings2_plot, labels = c( 'b', 'c' ), nrow = 2 )
right_panel

combined <- plot_grid( left_panel, right_panel, rel_widths = c( 1.0, 0.6 ) )
combined
```

# Session info
```{r session_information}
Sys.time()

getwd()
  
sessionInfo()
```
