==> markdowns/08_individual_genes.RMd <==
---
title: "Individual gene plots"
author: "Elisha Roberson"
date: "2020-10-14"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Project - Untreated + Skin + Muscle

```{r load_libraries}
library( here )
library( tidyverse )
library( cowplot )
library( reshape2 )
```

```{r source}
source( file = here::here( "src", "shared_r_functions.R" ) )
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	fig.path = paste0( figure_dir, "/" ),
	fig.keep = 'all',
	dpi = 300,
	fig.width = 11,
	fig.height = 8.5
)
```

```{r gene_annotation}
ensembl <- read_tsv( file = here::here( 'info', 'biomart_gene_transcript_map.txt.gz' ), col_names = TRUE ) %>%
  dplyr::rename( .data = ., gene_id = `Gene stable ID` ) %>%
  dplyr::rename( .data = ., symbol = `Gene name` ) %>%
  dplyr::rename( .data = ., gene_biotype = `Gene type` ) %>%
  select( .data = ., gene_id, symbol, gene_biotype ) %>%
  as.data.frame( . ) %>%
  unique( . )
rownames( ensembl ) = ensembl$gene_id

ensembl_gene_ids <- pull( ensembl, gene_id )

id_list <- ensembl$gene_id
names( id_list ) <- ensembl$symbol
```

```{r rgsm}
info <- read_csv( file = here::here( "info", "JDM_RNASeq_ReadGroup_Info.csv" ) ) %>%
  filter( Status %in% c( 'Control_PBMC', 'JDM_Untreated_PBMC', 'JDM_Inactive_PBMC' ) ) %>%
  select( -RGID ) %>%
  as.data.frame( . ) %>%
  unique( . ) %>%
  mutate( Status = case_when(
    str_detect( string = Status, pattern = "Control_PBMC" ) ~ "Control",
    str_detect( string = Status, pattern = "Untreated" ) ~ "Untreated",
    str_detect( string = Status, pattern = "Inactive" ) ~ "Inactive",
    TRUE ~ "Error"
  ) ) %>%
  mutate( Status = factor( Status, levels = c( 'Control', 'Untreated', 'Inactive' ) ) ) %>%
  as.data.frame( . )

rownames( info ) = info$RGSM
```

```{r rlog}
rlog <- read_tsv( file = here( 'results', 'unspliced_rlog_expression.tsv.gz' ), col_names = TRUE, col_types = cols( .default = 'd', gene_id = 'c' ) ) %>%
	as.data.frame( . ) %>%
	melt( . ) %>%
	dplyr::rename( RGSM = variable ) %>%
  filter( RGSM %in% rownames( info ) ) %>%
	dplyr::rename( rlog = value ) %>%
  mutate( sex = info[ RGSM, "Sex" ] ) %>%
	mutate( individual = info[ RGSM, "Individual" ] ) %>%
  mutate( Status = info[ RGSM, "Status" ] ) %>%
  mutate( Status = as.character( Status ) ) %>%
	mutate( Status = factor( Status, levels = c( "Control", "Untreated", "Inactive" ) ) ) %>%
	mutate( manual_x_position = case_when(
		Status == "Untreated" ~ 1.8,
		Status == "Inactive" ~ 3.0,
		TRUE ~ 1.0 ) )

```

```{r ggplot_gene_fxn}
return_ggplot_object <- function( current_gene, local_rlog = rlog, local_ensembl = ensembl, xlabel = "Status", ylabel = "Expression (rlog)" ) {
  current_symbol <- filter( local_ensembl, gene_id == current_gene ) %>%
    pull( symbol )
  
  #title_text <- paste0( current_gene, " (", current_symbol, ")\n" )
  #title_text <- expression( italic( paste( current_symbol ) ) )
  title_text <- current_symbol
  
  snip <- local_rlog %>%
    filter( gene_id == current_gene )
  
  for_lines = filter( snip, !is.na( individual ) ) %>%
  	mutate( Status = as.character( Status ) ) %>%
  	mutate( Status = factor( Status, levels = c( "Control", "Untreated", "Inactive" ) ) )
  
  ggplot( data = snip, mapping = aes( x = Status, y = rlog ) ) +
  	theme_bw() +
  	scale_colour_manual( values = colorBlindPalette ) +
  	stat_summary( fun.y = mean, fun.ymin = mean, fun.ymax = mean, geom = 'crossbar', width = 0.5 ) +
  	geom_point( size = 1.5, alpha = 0.8 ) +
  	geom_line( data = subset( snip, !is.na( individual ) ), mapping = aes( group = individual ) ) +
  	xlab( xlabel ) +
  	ylab( ylabel ) +
    ggtitle( title_text ) +
    theme( plot.title = element_text( face = 'italic' ) ) +
    theme( legend.position = "top" ) +
    gg_bigger_texts +
  	gg_center_title +
    expand_limits( y = 0 ) %>%
    return( . )
}
```

```{r ifi44l}
return_ggplot_object( 'ENSG00000137959' )
```

```{r quad_plot}
il1a_plot <-   return_ggplot_object( current_gene = unname( id_list[ 'IL1A' ] ), ylabel = "" )
il1b_plot <-   return_ggplot_object( current_gene = unname( id_list[ 'IL1B' ] ) )
il6_plot  <-   return_ggplot_object( current_gene = unname( id_list[ 'IL6'  ] ), xlabel = "", ylabel = "" ) + theme( axis.text.x = element_blank() )
cxcl10_plot <- return_ggplot_object( current_gene = unname( id_list[ 'CXCL10' ] ), xlabel = "" ) + theme( axis.text.x = element_blank() )

initial_cowplot <- plot_grid(
  cxcl10_plot + theme( legend.position = 'none' ) + ylab( "Expression (rlog)"),
  il6_plot    + theme( legend.position = 'none' ),
  il1b_plot   + theme( legend.position = 'none' ) + ylab( "Expression (rlog)"),
  il1a_plot   + theme( legend.position = 'none' ),
  nrow = 2,
  labels = c( 'a', 'b', 'c', 'd' )
)

#legend <- get_legend( il1a_plot )

#plot_grid( initial_cowplot, legend, rel_heights = c( 3, 0.3 ), nrow = 2 )
initial_cowplot
```

```{r complex_1_genes}
# MT-ND1;MT-ND2;MT-ND3;MT-ND4;MT-ND4L;MT-ND5
mtnd1_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND1' ] ), xlabel = "" ) + theme( axis.text.x = element_blank() )
mtnd2_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND2' ] ), xlabel = "", ylabel = "" ) + theme( axis.text.x = element_blank() )
mtnd3_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND3' ] ), xlabel = "", ylabel="" ) + theme( axis.text.x = element_blank() )
mtnd4_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND4' ] ), xlabel = "" ) + theme( axis.text = element_text( size = 15 ) )
mtnd4l_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND4L' ] ), ylabel = "" ) + theme( axis.text = element_text( size = 15 ) )
mtnd5_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MT-ND5' ] ), ylabel = "", xlabel = "" ) + theme( axis.text = element_text( size = 15 ) )

plot_grid( mtnd1_plot, 
           mtnd2_plot, 
           mtnd3_plot, 
           mtnd4_plot, 
           mtnd4l_plot, 
           mtnd5_plot, 
           labels = c( 'a', 'b', 'c', 'd', 'e', 'f' ),
           nrow = 2 )
```

```{r resolution_genes}
# RSAD2, CMPK2, IFI6, OTOF, IFI44, and MX1

cmpk2_plot <- return_ggplot_object( current_gene = unname( id_list[ 'CMPK2' ] ), xlabel = "" ) + theme( axis.text.x = element_blank() )

ifi6_plot <- return_ggplot_object( current_gene = unname( id_list[ 'IFI6' ] ), xlabel = "", ylabel = "" ) + theme( axis.text.x = element_blank() )

ifi44_plot <- return_ggplot_object( current_gene = unname( id_list[ 'IFI44' ] ), xlabel = "", ylabel = "" ) + theme( axis.text.x = element_blank() )

mx1_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MX1' ] ), xlabel = "" ) + theme( axis.text = element_text( size = 15 ) )

otof_plot <- return_ggplot_object( current_gene = unname( id_list[ 'OTOF' ] ), xlabel = "", ylabel = "" ) + theme( axis.text = element_text( size = 15 ) )

rsad2_plot <- return_ggplot_object( current_gene = unname( id_list[ 'RSAD2' ] ), xlabel = "", ylabel = "" ) + theme( axis.text = element_text( size = 15 ) )

plot_grid( cmpk2_plot, 
           ifi6_plot, 
           ifi44_plot, 
           mx1_plot, 
           otof_plot, 
           rsad2_plot, 
           labels = c( 'a', 'b', 'c', 'd', 'e', 'f' ),
           nrow = 2 )
```

```{r UntNotDE_2_DownInInactive}
# IFI16, MX2, OAS1


ifi16_plot <- return_ggplot_object( current_gene = unname( id_list[ 'IFI6' ] ), xlabel = "" ) + theme( axis.text = element_text( size = 15 ) )

mx2_plot <- return_ggplot_object( current_gene = unname( id_list[ 'MX2' ] ), xlabel = "", ylabel = "" ) + theme( axis.text = element_text( size = 15 ) )

oas1_plot <- return_ggplot_object( current_gene = unname( id_list[ 'OAS2' ] ), xlabel = "", ylabel = "" ) + theme( axis.text = element_text( size = 15 ) )

plot_grid( ifi16_plot, 
           mx2_plot, 
           oas1_plot, 
           labels = c( 'a', 'b', 'c' ),
           nrow = 1 )
```

# Session info
```{r session_information}
Sys.time()

getwd()
  
sessionInfo()
```
