==> markdowns/19_otof_rtqpcr.RMd <==
---
title: "OTOF confirmation"
author: "Elisha Roberson"
date: "2020-11-29"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Project - Untreated + Skin + Muscle

```{r libraries}
library( here )
library( reshape2 )
library( cowplot )
library( ggbeeswarm )
library( tidyverse )
```

```{r source}
source( file = here::here( "src", "shared_r_functions.R" ) )
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	fig.path = paste0( figure_dir, "/" ),
	fig.keep = 'all',
	dpi = 300,
	fig.width = 11,
	fig.height = 8.5
)
```

# Load data


```{r rtqpcr_input}
raw_data <- read_csv( file = here( "data", "OtOf_RTqPCR_tidy.csv" ) ) %>%
  filter( .data = ., RGSM != "Pachman_H20" )
```

```{r mean_sd}
mean_sd <- raw_data %>%
  dplyr::group_by( .data = ., RGSM, Target ) %>%
  dplyr::summarise(
    cq_n = length( which( !is.na( Cq ) ) ),
    cq_mean = mean( Cq, na.rm = TRUE ),
    cq_sd = sd( Cq, na.rm = TRUE )
  )
```

```{r dCq}
dCq <- mean_sd %>%
  dplyr::group_by( .data = ., RGSM ) %>%
  dplyr::summarize(
    dOtOf_mean = cq_mean[2] - cq_mean[1],
    dOtOf_sd = sqrt( sum( cq_sd^2, na.rm = TRUE ) )
  ) %>%
  mutate( .data = ., Status = case_when(
    str_detect( string = RGSM, pattern = "Control" ) ~ "Control",
    str_detect( string = RGSM, pattern = "Untreated" ) ~ "Untreated",
    str_detect( string = RGSM, pattern = "Inactive" ) ~ "Inactive",
    TRUE ~ "PROBLEM!"
  )) %>%
  mutate( .data = ., Status = factor( Status, levels = c( "Control", "Untreated", "Inactive" ) ) )
```

```{r anova}
tukey_output <- aov( formula = dOtOf_mean ~ Status, data = dCq ) %>%
  TukeyHSD( . )

tukey_output

untreated_v_control_pvalue <- tukey_output[[ 'Status' ]][ "Untreated-Control", "p adj" ]
inactive_v_untreated_pvalue <- tukey_output[[ 'Status' ]][ "Inactive-Untreated", "p adj" ]
```

```{r group_dCq}
group_dCq <- dCq %>%
  dplyr::group_by( .data = ., Status ) %>%
  dplyr::summarise(
    group_dOtOf_n = length( dOtOf_mean ),
    group_dOtOf_mean = mean( dOtOf_mean, na.rm = TRUE ),
    group_dOtOf_sd = sqrt( sum( dOtOf_sd^2, na.rm = TRUE ) )
  ) %>%
  mutate( .data = ., group_dOtOf_sd = group_dOtOf_sd / group_dOtOf_n ) # need to adjust the sd for the n in mean calculation

controls <- filter( .data = group_dCq, Status == "Control" )
control_dCq_mean <- controls$group_dOtOf_mean[1]
control_dCq_sd <- controls$group_dOtOf_sd[1]

group_dCq <- group_dCq %>%
  mutate( .data = ., mean_ddCq = group_dOtOf_mean - control_dCq_mean ) %>%
  mutate( .data = ., sd_ddCq = sqrt( sum( group_dOtOf_sd^2 + control_dCq_sd^2 ) ) ) %>%
  mutate( .data = ., mean_foldchange = 2^(-1 * mean_ddCq ) ) %>%
  mutate( .data = ., sd_foldchange = log(2) * mean_foldchange * sd_ddCq )

as.data.frame( group_dCq )
```

```{r RTqPCR_OtOf}
control_untreated_error_bar_y = 4.5 #16.5
inactive_untreated_error_bar_y = 3
y_offset = 0.4
asterisk_size = 6

ggplot( data = group_dCq, mapping = aes( x = Status, y = mean_foldchange ) ) +
  theme_bw() +
  gg_bigger_texts +
  xlab( "\nStatus" ) +
  ylab( "Relative fold-change\n" ) +
  geom_abline( intercept = 1.0, slope=0, lty=2 ) +
  geom_errorbar( aes( ymin = mean_foldchange - sd_foldchange, ymax = mean_foldchange + sd_foldchange ), width=0.2, position=position_dodge( 0.3 ), lwd=1.02 ) +
  geom_point( size = 3.5 ) +
  #ggtitle( "Otoferlin RT-qPCR foldchanges" ) +
  gg_center_title +
  geom_segment( aes( x = 2, xend = 3, y = inactive_untreated_error_bar_y, yend = inactive_untreated_error_bar_y ), size = 1.5 ) +
  geom_segment( aes( x = 1, xend = 2, y = control_untreated_error_bar_y, yend = control_untreated_error_bar_y ), size = 1.5 ) +
  geom_text( x = 1.5, y = control_untreated_error_bar_y + y_offset, label = paste0( "P = ", sprintf( "%.2e", untreated_v_control_pvalue ) ), size = asterisk_size, hjust = 0.5 ) +
  geom_text( x = 2.5, y = inactive_untreated_error_bar_y + y_offset, label = paste0( "P = ", sprintf( "%.2e", inactive_v_untreated_pvalue ) ), size = asterisk_size, hjust = 0.5 )
```

# Session info
```{r session_information}
Sys.time()

getwd()
  
sessionInfo()
```
